fix index error for cnv existing test

+++ tests/conftest.py
@@ -7606,7 +7606,8 @@ def discovered_apps_dr_workload_cnv(request, cnv_custom_storage_class):
     def teardown():
         try:
             instances[0].delete_workload(skip_resource_deletion_verification=True)
-            instances[1].delete_workload(shared_drpc_protection=True)
+            if request.node.name == "test_acm_kubevirt_using_shared_protection":
+                instances[1].delete_workload(shared_drpc_protection=True)
         except ResourceNotDeleted:
             raise ResourceNotDeleted("Workload deletion was unsuccessful")
